Update:Contact Page and Footer Components
